Use cases user

+++ spartan_games_app/src/main/java/group8/spartan_games_app/game/GameController.java
@@ -2,6 +2,8 @@ package group8.spartan_games_app.game;
 
 
 import java.io.IOException;
+
+import java.io.InputStream;
 import java.util.Base64;
 import java.util.List;
 
@@ -13,16 +15,15 @@ import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,6 +31,7 @@ import org.springframework.web.multipart.MultipartFile;
 import group8.spartan_games_app.review.Review;
 import group8.spartan_games_app.review.ReviewService;
 import group8.spartan_games_app.user.User;
+import group8.spartan_games_app.user.UserController;
 import group8.spartan_games_app.user.UserService;
 
 
@@ -106,6 +108,12 @@ public class GameController {
         model.addAttribute("game", game);
         //get reviews
         model.addAttribute("reviews", reviews);
+        //get dev info
+        User dev = userService.getUserById(game.getDevId());
+        model.addAttribute("dev", dev);
+
+        model.addAttribute("reviewList", reviewService.getReviewsByGameId(gameId));
+
         return "gamepage";
     }
 
@@ -141,6 +149,18 @@ public class GameController {
 
 
         return "search";
+
+=======
+    }
+
+    @GetMapping("/upload_game")
+    public String showUploadForm(Model model) {
+
+        userController.addUserAttributes(model); // Adds all the necessary attributes of the current user to the page
+        // It adds the attributes 'user' and 'thumbnailData', so don't repeat them after using this.
+
+        return "upload";
+
     }
 
     @GetMapping("/upload_game")
@@ -152,10 +172,6 @@ public class GameController {
         return "upload";
     }
 
-    @GetMapping("/upload_game")
-        public String showUploadForm(Model model) {
-        return "upload"; 
-}
 
     /**
      * Create a new Game entry.
@@ -177,10 +193,29 @@ public class GameController {
 
         service.addNewGame(title, description, devId, gameFile, thumbnailFile, gameFileName, thumbnailFileName);
 
-        return "redirect:/games/developer/" + devId;
+        return "redirect:/user/" + devId;
+    }
+
+
+    @GetMapping("/update/{gameId}")
+    public String showUpdateForm(@PathVariable int gameId, Model model) {
+    userController.addUserAttributes(model); 
+
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String username = authentication.getName();
+    User currentUser = userService.getUserByUsername(username);
+    Game game = service.getGameById(gameId);
+
+    System.out.println("CURRENT USER ID: " + currentUser.getUserId());
+
+    if (game.getDevId() != currentUser.getUserId()) {
+        return "403";
     }
     
+    model.addAttribute("game", game);
     
+    return "updategame"; 
+}
 
     /**
      * Performs the update
@@ -188,21 +223,22 @@ public class GameController {
      * @return
           * @throws IOException 
           */
-         @PutMapping("/update/{gameId}")
-         public Game updateGame(@PathVariable int gameId,
+    @PostMapping("/update/{gameId}")
+    public String updateGame(@PathVariable int gameId,
          @RequestParam(value = "title", required = false) String title,
          @RequestParam(value = "description", required = false) String description,
          @RequestParam(value = "gameFile", required = false) MultipartFile gameFile,
          @RequestParam(value = "thumbnailFile", required = false) MultipartFile thumbnailFile) throws IOException{
-        
+
+            
 
         service.updateGame(gameId, title, description, gameFile, thumbnailFile);
 
-        return service.getGameById(gameId);
+
+        return "redirect:/games/" + gameId;
+
+   
     }
-
-
-
 
     @GetMapping("/download/{gameId}")
     public ResponseEntity<Resource> downloadGameFile(@PathVariable int gameId) {
@@ -246,9 +282,46 @@ public class GameController {
      * @return the updated list of Games.
      */
 
-    @DeleteMapping("/delete/{gameId}")
-    public List<Game> deleteGameById(@PathVariable int gameId) {
-        service.deleteGameById(gameId);
-        return service.getAllGames();
+     @GetMapping("/delete/{gameId}")
+     public String deleteGameById(@PathVariable int gameId, Model model) {
+
+        userController.addUserAttributes(model); 
+        
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        User currentUser = userService.getUserByUsername(username);
+        Game game = service.getGameById(gameId);
+    
+        System.out.println("CURRENT USER ID: " + currentUser.getUserId());
+        if (game.getDevId() != currentUser.getUserId()) {
+            return "403";
+        }
+
+         int devId = game.getDevId();
+         service.deleteGameById(gameId);
+         
+         return "redirect:/user/" + devId;
+     }
+
+
+
+     //use for getting game thumbnails
+
+     @GetMapping("/{gameId}/thumbnail")
+    public ResponseEntity<byte[]> getUserThumbnail(@PathVariable int gameId) throws IOException {
+        Game game = service.getGameById(gameId);
+
+        if (game.getThumbnailData() == null) {
+            InputStream defaultImageStream = getClass().getResourceAsStream("/static/images/placeholder.jpg");
+            byte[] defaultImage = defaultImageStream.readAllBytes();
+            return ResponseEntity.ok()
+                    .contentType(MediaType.IMAGE_JPEG)
+                    .body(defaultImage);
+        }
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.IMAGE_PNG)
+                .body(game.getThumbnailData());
     }
+
 }
